+++ panel/src/components/projects/__tests__/project-card-grid.test.tsx
@@ -1,5 +1,26 @@
-import { describe, it, expect } from "vitest";
-import { render, screen } from "@testing-library/react";
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import { render, screen, fireEvent } from "@testing-library/react";
+
+const mockPush = vi.fn();
+vi.mock("next/navigation", () => ({
+  useRouter: () => ({ push: mockPush }),
+}));
+
+// The dialog itself (its own hooks/queries) is covered by
+// quick-edit-project-dialog.test.tsx; here we only care whether the name
+// click opens it with the right project id, not what it renders.
+vi.mock("../quick-edit-project-dialog", () => ({
+  QuickEditProjectDialog: ({
+    projectId,
+    open,
+  }: {
+    projectId: string;
+    open: boolean;
+  }) =>
+    open ? (
+      <div data-testid="quick-edit-dialog" data-project-id={projectId} />
+    ) : null,
+}));
 
 import { ProjectCardGrid } from "../project-card-grid";
 import { Team } from "@/types";
@@ -20,6 +41,10 @@ const project: ProjectSummary = {
 };
 
 describe("ProjectCardGrid", () => {
+  beforeEach(() => {
+    mockPush.mockClear();
+  });
+
   it("renders one card carrying the project's name, cell, tasks, token, status, and CI-watch badge", () => {
     render(<ProjectCardGrid projects={[project]} isLoading={false} />);
     expect(screen.getByText("RoboCo Core")).toBeInTheDocument();
@@ -53,5 +78,22 @@ describe("ProjectCardGrid", () => {
   it("does not show the empty state while loading", () => {
     render(<ProjectCardGrid projects={undefined} isLoading={true} />);
     expect(screen.queryByText("No projects found")).not.toBeInTheDocument();
+  });
+
+  it("the Edit action routes to the full settings page instead of opening a dialog", () => {
+    render(<ProjectCardGrid projects={[project]} isLoading={false} />);
+    fireEvent.click(screen.getByRole("button", { name: "Edit project" }));
+    expect(mockPush).toHaveBeenCalledWith("/projects/p1/settings");
+    expect(screen.queryByTestId("quick-edit-dialog")).not.toBeInTheDocument();
+  });
+
+  it("clicking the project name opens the slim quick-edit dialog instead of navigating", () => {
+    render(<ProjectCardGrid projects={[project]} isLoading={false} />);
+    fireEvent.click(screen.getByRole("button", { name: "RoboCo Core" }));
+    expect(mockPush).not.toHaveBeenCalled();
+    expect(screen.getByTestId("quick-edit-dialog")).toHaveAttribute(
+      "data-project-id",
+      "p1",
+    );
   });
 });
