+++ panel/src/components/projects/index.ts
@@ -2,5 +2,5 @@ export { ProjectTable } from "./project-table";
 export { ProjectCardGrid } from "./project-card-grid";
 export { ProjectFilters } from "./project-filters";
 export { CreateProjectDialog } from "./create-project-dialog";
-export { EditProjectDialog } from "./edit-project-dialog";
+export { QuickEditProjectDialog } from "./quick-edit-project-dialog";
 export { ProjectsView } from "./projects-view";
